Switch on relay by default
Today my remote controlled lamp has stopped working. I've took ladder and
pull it off. I found that MCU working well but relay refused to trigger.
I slightly tapped on relay's case and it began work as usual.
I think the problem is that before I set up relay to turn light on when it
triggered off and when relay is up, lamp is turned off. As most time lamp is
off thats mean that relay is switched on. So I've got some problem with
mechanical  part.
Now lamp operates in more natural mode: When relay is on, lamp is on and vise
versa. That requred additional screwdriver work and single line change in MCU
firmware. Hope this will eliminate problem wih relay.